requirements.txt
wheel
Django==2.0
pyaml
pytz
psycopg2
django-guardian
